% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_storage_system_arg}
\alias{check_storage_system_arg}
\title{Check if storage system argument is valid}
\usage{
check_storage_system_arg(storage_system)
}
\arguments{
\item{storage_system}{input argument to check}
}
\value{
None
}
\description{
Check if storage system argument is valid
}
