% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_rsurveymethods_file_name}
\alias{create_rsurveymethods_file_name}
\title{Appends Rsurveymethods and version in file name}
\usage{
create_rsurveymethods_file_name(input_data_path)
}
\arguments{
\item{input_data_path}{full path name}
}
\value{
filename
}
\description{
Appends Rsurveymethods and version in file name
}
