% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{download_file_from_s3}
\alias{download_file_from_s3}
\title{Download file from AWS S3 locally}
\usage{
download_file_from_s3(input_full_path_s3, local_path)
}
\arguments{
\item{input_full_path_s3}{AWS S3 path where data are stored}

\item{local_path}{path to local storage}
}
\value{
None
}
\description{
Download file from AWS S3 locally
}
