% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_results.R
\name{export_results}
\alias{export_results}
\title{Export merged survey results to CSV}
\usage{
export_results(input_data_with_counts, size_band_estimates, output_file)
}
\arguments{
\item{input_data_with_counts}{Data frame with pre-processed survey data}

\item{size_band_estimates}{Data frame with size-band level estimates}

\item{output_file}{Path to output CSV file}
}
\value{
None. Writes merged data to CSV.
}
\description{
Export merged survey results to CSV
}
