% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_path}
\alias{format_path}
\title{Format file path by adding s3 prefix if needed}
\usage{
format_path(config, path, platform = c("s3", "network"))
}
\arguments{
\item{config}{configuration list}

\item{path}{original file path}

\item{platform}{storage platform, either "s3" or "network"}
}
\value{
formatted file path
}
\description{
Format file path by adding s3 prefix if needed
}
