% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_se_for_publication}
\alias{format_se_for_publication}
\title{format standard errors file for publication}
\usage{
format_se_for_publication(df, selected_period = "")
}
\arguments{
\item{df}{dataframe to format for standard errors publication}

\item{selected_period}{optional, if provided only that period will be processed YYYYMM format}
}
\value{
formatted dataframe
}
\description{
format standard errors file for publication
}
