% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_processing.R
\name{pre_process_data}
\alias{pre_process_data}
\title{Pre-process survey input data by merging with population counts and
calculating derived variables.}
\usage{
pre_process_data(input_data, population_counts)
}
\arguments{
\item{input_data}{Data frame with survey input data. Must contain columns:
period, cell, aweight, gweight, value, outlier_weight.}

\item{population_counts}{Data frame with population counts. Must contain
columns: period, cell.}
}
\value{
Data frame with merged data and new columns: extcalweights,
winsorised_value.
}
\description{
Pre-process survey input data by merging with population counts and
calculating derived variables.
}
\examples{
# input_data <- read.csv("artifact/input/input_regen_test_data.csv")
# population_counts <- read.csv("artifact/input/regen_population_counts_test.csv")
# processed <- pre_process_data(input_data, population_counts)

}
