% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_functions.R
\name{read_input_data}
\alias{read_input_data}
\title{Read input data}
\usage{
read_input_data(input_path)
}
\arguments{
\item{input_path}{Filepath to input data.}
}
\value{
DataFrame.
}
\description{
Read input data
}
