% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regenesses_estimation.R
\name{regenesses_estimation}
\alias{regenesses_estimation}
\title{Perform domain estimation using calibrated survey design}
\usage{
regenesses_estimation(input_data_with_counts)
}
\arguments{
\item{input_data_with_counts}{Data frame with pre-processed survey data}
}
\value{
A list with total and size-band estimates
}
\description{
Perform domain estimation using calibrated survey design
}
