% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{run_hdfs}
\alias{run_hdfs}
\title{Run an hdfs command as hdfs dfs arg1 arg2}
\usage{
run_hdfs(...)
}
\arguments{
\item{...}{arguments to run}
}
\value{
r command status
}
\description{
Run an hdfs command as hdfs dfs arg1 arg2
}
