% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{run_regenesess}
\alias{run_regenesess}
\title{Function to run the regenesess main process}
\usage{
run_regenesess(
  storage_system = c("network", "s3"),
  input_data_path,
  population_counts_path,
  output_path,
  run_id,
  debug_mode,
  selected_period = ""
)
}
\arguments{
\item{storage_system}{accepts local or s3}

\item{input_data_path}{file path of main dataframe}

\item{population_counts_path}{file path of population counts}

\item{output_path}{file path to save results}

\item{run_id}{unique identifier for the run}

\item{debug_mode}{boolean to indicate if debug mode is on, if true will output additional file}

\item{selected_period}{optional, if provided only that period will be processed YYYYMM format}
}
\description{
Function to run the regenesess main process
}
