% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{upload_file_to_s3}
\alias{upload_file_to_s3}
\title{Upload file to AWS S3}
\usage{
upload_file_to_s3(local_full_path, save_path_s3)
}
\arguments{
\item{local_full_path}{path to file}

\item{save_path_s3}{path to S3}
}
\value{
None
}
\description{
Upload file to AWS S3
}
