% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_file_name}
\alias{format_file_name}
\title{Format file name by appending run_id}
\usage{
format_file_name(config, file_name, run_id, platform = "s3")
}
\arguments{
\item{file_name}{original file name}

\item{run_id}{unique identifier for the run}
}
\value{
formatted file name
}
\description{
Format file name by appending run_id
}
