% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_file_name}
\alias{format_file_name}
\title{Format file name by appending run_id and adding s3 prefix if needed}
\usage{
format_file_name(config, path, file_name, platform)
}
\arguments{
\item{config}{configuration list}

\item{path}{original file path}

\item{file_name}{original file name}

\item{platform}{storage platform, either "s3" or "network"}
}
\value{
formatted file name
}
\description{
Format file name by appending run_id and adding s3 prefix if needed
}
