% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_se_for_publication}
\alias{format_se_for_publication}
\title{format standard errors file for publication}
\usage{
format_se_for_publication(
  df,
  storage_system,
  output_path,
  selected_period = ""
)
}
\arguments{
\item{df}{dataframe to format for standard errors publication}

\item{storage_system}{should be local or s3}

\item{output_path}{directory to output formatted df}

\item{selected_period}{optional, if provided only that period will be processed YYYYMM format}
}
\value{
None
}
\description{
format standard errors file for publication
}
