% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{main}
\alias{main}
\title{Main function to run pipeline process}
\usage{
main(config_path, platform = "s3")
}
\arguments{
\item{config_path}{path to configuration json file}
}
\value{
None
}
\description{
Main function to run pipeline process
}
