% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{main}
\alias{main}
\title{Main Function to run the project}
\usage{
main(input_data_path, population_counts_path, output_path)
}
\arguments{
\item{input_data_path}{file path of main dataframe}

\item{population_counts_path}{file path of population counts}

\item{output_path}{file path to save results}
}
\description{
Main Function to run the project
}
