% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_csv_wrapper}
\alias{read_csv_wrapper}
\title{Wrapper csv reader for local or s3 data storage}
\usage{
read_csv_wrapper(storage_system, input_data_path)
}
\arguments{
\item{storage_system}{accepts local or s3}

\item{input_data_path}{full path of data}
}
\value{
df dataframe
}
\description{
Wrapper csv reader for local or s3 data storage
}
